dev-requirements.txt
-r requirements.txt
flake8==3.5.0
google-cloud-firestore==1.2.0
google-cloud-datastore==1.8.0
google-cloud-tasks==1.1.0
pytest==3.8.2
pytest-cov==2.6.0
